minor
Former-commit-id: 7931dfea4f83c25d50d5ead0472f82ec28928779
Former-commit-id: aa0ba76622100e6c38d064ac5727b691d5bdefe3
Former-commit-id: 69aa7b047726c8f2746a948c9e9d79aebcb15b6f
Former-commit-id: 03a0fa577a062935ff359ec96f69d3dbbb8301f0 [formerly 38cd6fd257f5d2e85241454bbafce82acfe1653f]
Former-commit-id: 9387f35b38ea844700fc04ab8651d67be7f62420

+++ tests/ccsd/ccsd_t/ccsd_t_doubles_gpu.hpp
@@ -42,7 +42,7 @@ void ccsd_t_doubles_gpu(ExecutionContext& ec,
                    Tensor<T>& d_t2, //d_a
                    Tensor<T>& d_v2, //d_b
                    std::vector<T>& p_evl_sorted,
-                   std::vector<int>& k_range,
+                   std::vector<size_t>& k_range,
                    int t_h1b, int t_h2b, int t_h3b,
                    int t_p4b, int t_p5b, int t_p6b,
                    int usedevice=1) {
